Phase 2 - V4
Changement accuracy , mettre le calcul dans WellClassified

+++ Project/Project/Biblio.h
@@ -44,7 +44,7 @@ typedef struct classe Classe;
 struct classe {
 	int nbWellClassified;
 	int nbTested;
-	int accuracy;
+	double accuracy;
 	int remplacements[NB_CLASSE_MAX];
 };
